First attempt at s2 c13
The encoding/decoding of the query string is handled easily enough but
Rust's HashMap type is initialized by default with a random component
added to the hasher. This is great for preventing certain kinds of
attacks on the software but really sad making for people playing
attacker. I guess I'll have to pivot my approach to use a different
structure.